move getBrowserStorageEffectForKey to a different module to avoid circular dependency
